removed old extended files
