--- src/main/java/com/contrastsecurity/core/extended/EventDetails.java
@@ -1,47 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017 Contrast Security.
- * All rights reserved. 
- * 
- * This program and the accompanying materials are made available under 
- * the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation; either version 3 of the License.
- * 
- * The terms of the GNU GPL version 3 which accompanies this distribution
- * and is available at https://www.gnu.org/licenses/gpl-3.0.en.html
- * 
- * Contributors:
- *     Contrast Security - initial API and implementation
- *******************************************************************************/
-package com.contrastsecurity.core.extended;
-
-import java.util.List;
-
-public class EventDetails {
-	private boolean success;
-	private List<String> messages;
-	private Event event;
-
-	public void setSuccess(boolean success) {
-		this.success = success;
-	}
-
-	public boolean getSuccess() {
-		return this.success;
-	}
-
-	public void setMessages(List<String> messages) {
-		this.messages = messages;
-	}
-
-	public List<String> getMessages() {
-		return this.messages;
-	}
-
-	public void setEvent(Event event) {
-		this.event = event;
-	}
-
-	public Event getEvent() {
-		return this.event;
-	}
-}
