--- src/main/java/com/contrastsecurity/core/extended/EventModel.java
@@ -1,15 +1,0 @@
-package com.contrastsecurity.core.extended;
-
-public class EventModel {
-	
-	Object parent;
-
-	public Object getParent() {
-		return parent;
-	}
-
-	public void setParent(Object parent) {
-		this.parent = parent;
-	}
-
-}
