--- src/main/java/com/contrastsecurity/core/extended/FilterResource.java
@@ -1,34 +1,0 @@
-package com.contrastsecurity.core.extended;
-
-import java.util.List;
-
-public class FilterResource {
-
-    private boolean success;
-    private List<String> messages;
-    private List<Filter> filters;
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public void setSuccess(boolean success) {
-        this.success = success;
-    }
-
-    public List<String> getMessages() {
-        return messages;
-    }
-
-    public void setMessages(List<String> messages) {
-        this.messages = messages;
-    }
-
-    public List<Filter> getFilters() {
-        return filters;
-    }
-
-    public void setFilters(List<Filter> filters) {
-        this.filters = filters;
-    }
-}
