+++ src/main/java/com/contrastsecurity/ui/com/contrastsecurity/ui/toolwindow/ContrastToolWindowFactory.java
@@ -20,7 +20,6 @@ import com.contrastsecurity.config.ContrastUtil;
 import com.contrastsecurity.core.Constants;
 import com.contrastsecurity.core.cache.ContrastCache;
 import com.contrastsecurity.core.cache.Key;
-import com.contrastsecurity.core.extended.StatusRequest;
 import com.contrastsecurity.core.internal.preferences.OrganizationConfig;
 import com.contrastsecurity.exceptions.UnauthorizedException;
 import com.contrastsecurity.http.TraceFilterForm;
@@ -37,6 +36,7 @@ import com.contrastsecurity.models.RecommendationResponse;
 import com.contrastsecurity.models.Risk;
 import com.contrastsecurity.models.RuleReferences;
 import com.contrastsecurity.models.Server;
+import com.contrastsecurity.models.StatusRequest;
 import com.contrastsecurity.models.StoryResponse;
 import com.contrastsecurity.models.Tag;
 import com.contrastsecurity.models.Tags;
