+++ src/main/resources/contrast.properties
@@ -1,0 +1,1 @@
+version=2.10.0-SNAPSHOT
